+++ CHANGELOG.md
@@ -6,6 +6,7 @@ format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 #### Added
 #### Changed
+- Remove copyright warning, as we now have permission from GOBLINKO!
 - Fiddle with expansion order (put Hand of Doom KS Exclusives first, as
   it was, chronologically).